+++ packages/cf-deploy-config-writer/src/cf-writer/base-config.ts
@@ -14,7 +14,7 @@ import {
 } from '../utils';
 import LoggerHelper from '../logger-helper';
 import { t } from '../i18n';
-import { MtaConfig, createMTA, addMtaDeployParameters, addMtaBuildParams } from '../mta-config';
+import { MtaConfig, createMTA, addMtaDeployParameters } from '../mta-config';
 import { type Logger } from '@sap-ux/logger';
 import { type CFBaseConfig, RouterModuleType, type MTABaseConfig } from '../types';
 
@@ -56,7 +56,6 @@ async function addRoutingConfig(config: CFBaseConfig, fs: Editor): Promise<void>
         await mtaConfigInstance.addRoutingModules(true);
     }
     await addMtaDeployParameters(mtaConfigInstance);
-    await addMtaBuildParams(mtaConfigInstance);
     await mtaConfigInstance.save();
 }