+++ packages/inquirer-common/jest.config.js
@@ -1,4 +1,5 @@
 const config = require('../../jest.base');
+config.setupFilesAfterEnv = ['jest-extended/all'];
 config.snapshotFormat = {
     escapeString: false,
     printBasicPrototype: false