+++ packages/fiori-annotation-api/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @sap-ux/fiori-annotation-api
 
+## 0.3.0
+
+### Minor Changes
+
+-   09a58bb: [BREAKING CHANGE] Changed types for dynamic expressions to be compatible with the changes in @sap-ux/vocabulary-types (https://github.com/SAP/open-ux-odata/pull/844).
+
+## 0.2.17
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.28.7
+-   @sap-ux/cds-odata-annotation-converter@0.3.9
+
+## 0.2.16
+
+### Patch Changes
+
+-   0235973: Update @sap/ux-cds-compiler-facade dependency
+-   Updated dependencies [0235973]
+    -   @sap-ux/cds-odata-annotation-converter@0.3.9
+
 ## 0.2.15
 
 ### Patch Changes