+++ packages/ui-components/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/ui-components
 
+## 1.21.2
+
+### Patch Changes
+
+-   2a72ad2: chore - Fix audit issues
+
+## 1.21.1
+
+### Patch Changes
+
+-   32191dd: UITranslationInput. Screen readers read the resolved translation tooltip/title of UITranslationInput as part of the screen reader speech, displaying it as 'Value: ... Translation: ...'.
+
 ## 1.21.0
 
 ### Minor Changes